requirements.txt
Flask==3.0.3
Flask-BasicAuth==0.2.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
python-decouple==3.8
requests==2.32.3
ultralytics==8.2.35
psycopg2==2.9.9
gunicorn
dill
